Begin on a shitty queue implementation.
The `struct`s here are more or less inspired by the ones in
`crossbeam-epoch`. While the use cases for the two crates probably are
different, I think it would be smart to reuse the `struct`s there as much
as possible, and eventually remove stuff if we find out that we do not
need it.

Eg, `crossbeam-epoch` have their own `Atomic`, which is an atomic
pointer, which points to some data that is shared, as opposed to
`Owned`, which seems to basically be `Box`. We probably can use
something like this, so we do not have to keep converting between
`std::sync::atomic::AtomicPointer` and `usize`, for `null` checks;
with our own atomic pointer we can simply add stuff like that in the
`impl`.