Add concurrent test for list. Fails.
Make List::contains, which is probably OK.
List::remove is broken. See the hash table paper for a correct version.
We need to tag pointers - is there a better way? This is starting to
look like some reclamaition scheme. Maybe if we do not care about this
we don't have to tag the pointers?