Add `Pin`s back to `ebr`, and implement epoch logic
The threads read the global epoch, and every once in a while tries to
update the global epoch. Garbage is not yet handeled, but it is only a
matter of having lists to save it in, and actually do the collecting.
It _should_ be straight forward.

We should also try to find a more streamlined way of testing the memory
usage than just checking `htop`. We could also think about having some
abstraction around node boxing, so that we can confirm that all nodes
are reclaimed properly.

An idea is to simply have `[de]allocate_node`, and have a concurrent
list inside which holds some metadata (or even just a `Mutex<HashMap>`).
We could use `#cfg`s to only test for this if we really want, or in
debug builds or something.