Make ebr::queue compile
We've begun on the garbage handling, but there are some problems;
primarily this approach will not work, since every node in a list will
be put into another list, which also needs to be garbage collected, so
its nodes will be put into the same list, which will make a new node,
etc.

`crossbeam-epoch` have a `Bag` thing, which groups garbage, and flushes
it when it is full. We should probably do something like this.
Alternatively, we can do something simpler, like wrapping the list in a
Mutex, but this will probably hurt performance.