Begin on actually freeing garbage
We make a global list of (epoch, Garbage), which we free when we increment
the epoch. However, `Garbage` is currently a `FnOnce`, since we do not know
the type etc. of the data we are freeing.  Unfortunately, actually calling
the `FnOnce` wans't as simple as I had hoped, so we do not call the
closure yet, meaning we shouldn't really free any data;
however the closure frees the captured data when it gets freed (when
`Garbage` goes out of scope - confirmed by `mem::forget`).

Maybe letting the closure do the freeing is alright for now, since we
don't want to do any other things in the closure anyways. This also
seems like a fair place to stop, seing as we're only testing the
performance of EBR, and not making a full system here.

In any case, we segfaul, so something is wrong.