Debug the segfault
What we've found out so far: the address of the closure that when fred
is causing the segfault has been used for a closure once before. So, it
looked like a double free, but there are two `Garbage` objects created
at that address, so the memory is reused. The next step is to find out
if this is a problem (not very likely), or if it is the garbage in the
closure that is the problem.