Fix segfault, and confirm payload free.
The segfault came from a double free, which was caused since certain
memory were freed once, and then again when the node holding the bag it
was in the first time got freed. By using `ManuallyDrop`, we avoid
freeing it the second time. By running `queue::test::memory_usage` and
watching `htop`, we see that the payload memory does get freed, as it
should. I am not sure if there are any leaks, but it seems like there
are none.

Concurrent memory usage test is not done, so there may still be bugs
here.