Move ManuallyDrop to queue::Node
The queue should not `drop` the elements of the queue - this is the job
of the user of the queue. Eg, when an element is popped from the queue
the value returned will be dropped when it goes out of scope. This is
the case for all values in the queue, even the one we use in the memory
reclamation itself.

We also add some tests to confirm that nothing is dropped when we don't
expect it, that all structs are dropped, and that they are dropped
exactly once.