Fix garbage epoch tagging, and check
The `>= 5` was just a placeholder, to make sure that previous bugs
weren't off-by-one errors. The `add_garbage_bag` change is due to a
fundamental misunderstanding with EBR - we are not supposed to tag
garbage with the epoch that the thread is in, but the global epoch
at the time of the `push`ing. This matters, or else the following could
happen: consider two threads, A and B.

(e)   A wants to retire X
(e)   B increments the epoch
(e+1) B gets a reference to A
(e)   A pushes (e, X) to the garbage list
(e)   A increments the epoch to e+2
      Now we are in e+2, and X is tagged with e,
      but B still has a reference to X.

By using the global epoch, this cannot happen, since X is unlinked
before reading the global epoch. Now A will either read e, in which case
B cannot get a reference to X, or it will read e+1, but then all is OK.