Implement `queue::drop`, and add comments.
We need `queue::drop` in order to drop items in the queue when the queue
itself is dropped. A todo is to make sure that the queue is only dropped
from one thread.

We are still investigating the illegal read in `pop_if`, and we have
valgrind backtraces for it. However, it doesnt really make any sense.