Remove custom ordering stuff in `hp/atomic.rs`
We will (probably) not use this, so we'll remove it. Maybe we want to
actually have different orderings for succ/fail paths, but I think we
rather want to just use `SeqCst` for all orderings; this is probably
fine for us, since ARMv8 have CS load/store instructions (although we
don't know how they perform on our CPU, compared to regular
load/stores).