Remove `hp::queue::{push,pop}_node` functions.
They were buggy, and not really needed. Also make a new function that
takes a closure which is passed the HP before returning. This makes it
possible to either push it to the hp queue or spin, such that we don't
need to chunk up the HPs in blocks, as we do with EBR. Chunking would be
possible, but tiresome, as we would need to mark all HPs in a block or
something, so that we don't `scan` all of them every check for when the
block is safe to free.

`hp::queue::test::stress_test` runs fine.