Begin on thread cleanup of global state
This is a problem for both EBR and HP, and soudlv'e probably be done in
two commits, but oh well.

EBR and HP have global state: both have a list of thread entries which
is scanned regularily. A thread inserts the entry in the `thread_local!`
macro, but the entry is never removed. This causes scans of threads that
are long gone, which may take a lot of time. We saw this in the
benchmarks, where some tests that originally were 13ms per test got up
to 100ms.

The idea is simple: have a type that is thread local, and impl Drop on
it, and remove the thread entry in the Drop impl. However, there are
some complications here, such that we cannot use MR while removing the
entry from the list, since we are destroying the local state. This
causes weird API changes. This commit started messing around with the
return value of `hp::list::List::remove`, from `Ptr<T>` to
`Ptr<Node<T>>`, since EBR does that. We should rethink what we need, and
fix both modules before working more on the benchmarks. We also have to
`flush` the local buffer, so that we are not leaking any memory.

This might cause us to have to manage destructors of types very
carefully. An idea for doing this is to have nested structs, since we
get certain guarantees on the order of the `Drop`s there.

I also messed around with `Condvar` vs `Barrier`, but there were no
noticable difference in the benchmark, which is good.