Continue on thread cleanup logic.
We still have the same problem: thread local stuff needs itself when
cleaning up, but that is not allowed. One solution could be to have
methods in which we pass the thread local stuff around, although this is
very cumbersome. Then the other regular methods can just get it, and
pass it around.