`benches/queue2` is now a soemwhat working benchmark.
The "benchmark" is now just a binary, which calls `bench` stuff. The
timings we get back is very varying, but the numbers are alright,
meaning they are pretty much the what we normally get. Single threaded
performance is still better than multiple threads, even with just 2 of
them. We should probably check recent literature for what we should test
in order to compare HP and EBR.

We should probably fix the thread number stuff here, so that we don't
have to make a new binary for each new thread # check, as that's pretty
dumb. It seems the thread cleanup hacks are getting worse and worse by
the day; maybe we should reconsider cleaning up instead?

We should also make a HP benchmark similar to this, and look a little
closer at the performance difference.