Write a thread benchmarking system
The newer system spawns threads and handles communication and
synchronization, so that the client (the benchmarks) only need to supply
a shared state and the function that is to be ran. We are getting the
same numbers as the previous system, but with _a lot_ less boilerplate
for the benchmarks.