Rework the whole benchmarking system.
We now have a new crate, `benchmark-runner`. This crate contains all
benchmarks, as well as a simple CLI for handling the options to it. Was
is really neccessarry? Probably maybe not. But now we get outputted all
the data we need, and it is rather simple to add new benchmarks. In
addition we have a format on the data files now, so it should be
possible to create scripts for generating the graphs. The graph gen
should be done in the report repository.