Fix a use after free
`remove_thead_local` cannot use the regular remove function, as we
remove, *and free* the node with the thread local thing. This causes the
hp::deregsiter to lookup the marker, and read garbage, since we do not
change the thread local pointer. Instead, we must simply return the
entire node, and let HP deal with when to release it. We do this by
scanning, since this is only on thread shutdown.

valgrind was ran without any errors, and cargo run also didn't crash,
so it seems we are in the green.